Move cbook._check_shape() to _api.check_shape()
