+++ lib/matplotlib/tests/test_cbook.py
@@ -1,6 +1,5 @@
 import itertools
 import pickle
-import re
 
 from weakref import ref
 from unittest.mock import patch, Mock
@@ -673,22 +672,6 @@ def test_array_patch_perimeters():
         for rstride, cstride in itertools.product(divisors(rows - 1),
                                                   divisors(cols - 1)):
             check(x, rstride=rstride, cstride=cstride)
-
-
-@pytest.mark.parametrize('target,test_shape',
-                         [((None, ), (1, 3)),
-                          ((None, 3), (1,)),
-                          ((None, 3), (1, 2)),
-                          ((1, 5), (1, 9)),
-                          ((None, 2, None), (1, 3, 1))
-                          ])
-def test_check_shape(target, test_shape):
-    error_pattern = (f"^'aardvark' must be {len(target)}D.*" +
-                     re.escape(f'has shape {test_shape}'))
-    data = np.zeros(test_shape)
-    with pytest.raises(ValueError,
-                       match=error_pattern):
-        cbook._check_shape(target, aardvark=data)
 
 
 def test_setattr_cm():
